--- BE/airbnb/src/main/java/team07/airbnb/entity/Image.java
@@ -1,31 +1,0 @@
-package team07.airbnb.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@EqualsAndHashCode(of = "IMAGE_ID")
-public class Image {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "IMAGE_ID")
-    private Long id;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "ROOM_ID")
-    private Room room;
-    private String imageUrl;
-    @Enumerated(value = EnumType.STRING)
-    private ImageType imageType;
-
-}
