--- BE/airbnb/src/main/java/team07/airbnb/entity/User.java
@@ -1,21 +1,0 @@
-package team07.airbnb.entity;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-@Entity
-@EqualsAndHashCode(of = "USER_ID")
-@NoArgsConstructor
-public class User {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;
-    private String email;
-    private String nickName;
-    private String phone;
-
-}
