+++ BE/airbnb/src/main/java/team07/airbnb/image/Image.java
@@ -1,7 +1,10 @@
-package team07.airbnb.entity;
+package team07.airbnb.image;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -9,20 +12,21 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import team07.airbnb.room.Room;
 
 @Entity
-@EqualsAndHashCode(of = "WISHLIST_ID")
 @NoArgsConstructor
-public class Wishlist {
+@EqualsAndHashCode(of = "id")
+public class Image {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "WISHLIST_ID")
+    @Column(name = "IMAGE_ID")
     private Long id;
-    @ManyToOne
-    @JoinColumn(name = "USER_ID")
-    private User user;
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "ROOM_ID")
     private Room room;
+    private String imageUrl;
+    @Enumerated(value = EnumType.STRING)
+    private ImageType imageType;
 
 }
