+++ BE/airbnb/src/main/java/team07/airbnb/wishlist/Wishlist.java
@@ -1,6 +1,5 @@
-package team07.airbnb.entity;
+package team07.airbnb.wishlist;
 
-import java.time.LocalDate;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -10,23 +9,22 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import team07.airbnb.room.Room;
+import team07.airbnb.user.User;
 
 @Entity
+@EqualsAndHashCode(of = "id")
 @NoArgsConstructor
-@EqualsAndHashCode(of = "REVIEW_ID")
-public class Review {
+public class Wishlist {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "REVIEW_ID")
+    @Column(name = "WISHLIST_ID")
     private Long id;
-    @ManyToOne
-    @JoinColumn(name = "ROOM_ID")
-    private Room room;
     @ManyToOne
     @JoinColumn(name = "USER_ID")
     private User user;
+    @ManyToOne
+    @JoinColumn(name = "ROOM_ID")
+    private Room room;
 
-    private int starRating;
-    private String content;
-    private LocalDate writeDate;
 }
